added dashboard api as well as other changes

+++ src/main/java/com/raj/jewellers/respository/JewelleryItemRepository.java
@@ -7,6 +7,6 @@ import com.raj.jewellers.entity.JewelleryItemEntity;
 
 @Repository
 public interface JewelleryItemRepository
-        extends JpaRepository<JewelleryItemEntity, Integer> {
+        extends JpaRepository<JewelleryItemEntity, Long> {
 
 }
